requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.8
django-filter==21.1
gunicorn==20.1.0
psycopg2==2.9.1
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
xlwt==1.3.0
